requirements.txt
ansible-base==2.10.8
apache-libcloud==3.2.0
apturl==0.5.2
argcomplete==1.8.1
Babel==2.8.0
bcrypt==3.2.0
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
build==0.10.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.1.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
commonmark==0.9.1
cryptography==39.0.0
cupshelpers==1.0
db-dtypes==1.0.5
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
dnspython==2.1.0
docutils==0.19
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.63.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.6.0
google-cloud-bigquery-storage==2.19.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.56.4
grpcio==1.51.3
grpcio-status==1.51.3
httplib2==0.20.4
idna==3.4
importlib-metadata==6.0.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.0.3
jmespath==0.10.0
keyring==23.13.1
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==9.0.0
netaddr==0.8.0
netifaces==0.11.0
ntlm-auth==1.4.0
numpy==1.23.3
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pandas==1.5.0
pandas-gbq==0.19.1
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.9.6
platformdirs==2.5.1
proto-plus==1.22.2
protobuf==4.21.7
psycopg2==2.9.2
ptyprocess==0.7.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycryptodomex==3.11.0
pycups==2.0.1
pydata-google-auth==1.7.0
Pygments==2.14.0
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyproject_hooks==1.0.0
pyRFC3339==1.1
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
pytz==2022.2.1
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
readme-renderer==37.3
reportlab==3.6.8
requests==2.28.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rfc3986==2.0.0
rich==13.1.0
rsa==4.9
SecretStorage==3.3.3
selinux==3.3
simplejson==3.17.6
six==1.16.0
systemd-python==234
tomli==2.0.1
twine==4.0.2
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.12
usb-creator==0.3.7
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xkit==0.0.0
xmltodict==0.12.0
zipp==3.11.0
